Problem: everything is tied together
Solution: split the gonix/pipe part to github.com/gomoni/gio and
github.com/gomoni/unix providing a minimal base for others to build on
top of. Port the gonix on top of gio/unix.

Remove wasm code from repo - while useful, it does not belong to native
Go utilities and brings the dependency on wazero runtime for everything.

Drop the shell colon parsing, this may be replaced by an interesting
https://github.com/mvdan/sh project.